# Faster circle and line algos