I have added Midpoint Circle Algorithm and Bresenham's algorithm to draw circles and lines faster.

I created some benchmarks and here are the results:\

BenchmarkBuffer_Line-8         	18535812	        61.44 ns/op
BenchmarkBuffer_LineFast-8     	35687457	        34.03 ns/op
BenchmarkBuffer_Circle-8       	 1000000	      1028 ns/op
BenchmarkBuffer_CircleFast-8   	20413906	        58.19 ns/op